Move image.datasets -> image.volume_space_management
